Rename settings files. Initialize eb app. Create .ebextensions command.
