+++ .ebextensions/01_packages.config
@@ -1,0 +1,4 @@
+packages:
+  yum:
+    git: []
+    postgresql93-devel: []
