+++ .gitignore
@@ -13,3 +13,8 @@ client/.cache/
 client/dist/
 client/node_modules/
 client/build/
+
+# Elastic Beanstalk Files
+.elasticbeanstalk/*
+!.elasticbeanstalk/*.cfg.yml
+!.elasticbeanstalk/*.global.yml
